moved generate children to utils folder
isolated grayed out css function
