+++ src/components/HelloWorld.vue
@@ -18,6 +18,7 @@
 import { tree } from 'vued3tree';
 import ServiceBrances from '@/static/ServiceBranches';
 import LevelTraverseTree from '@/utils/LevelTraverseTree';
+const generateChildren = require('../utils/GenerateChildren').generateChildren;
 
 export default {
   components: {
@@ -29,34 +30,7 @@ export default {
     return {
       tree: {
         name: ServiceBrances[Math.floor(Math.random()*ServiceBrances.length)],
-        children:(()=>{
-          const numberOfChildren = Math.floor(Math.random() * 5) + 1;
-          const children = [];
-          for(let i = 0; i < numberOfChildren; i+=1){
-            children.push({name: ServiceBrances[Math.floor(Math.random()*ServiceBrances.length)]})
-          }
-          for(let i = 0; i < children.length; i+=1){
-            const child = children[i];
-            const grandChildren = [];
-            const numberOfGrandChildren = Math.floor(Math.random() * 5);
-            for(let j = 0; j < numberOfGrandChildren; j+=1){
-              grandChildren.push({name: ServiceBrances[Math.floor(Math.random()*ServiceBrances.length)]})
-            }
-            child.children = grandChildren;
-          }
-          for(let i = 0; i < children.length; i+=1){
-            for(let j = 0; j < children[i].children.length; j+=1){
-              const grandChild = children[i].children[j];
-              const numberOfGreatGrandChildren = Math.floor(Math.random() * 5);
-              const greatGrandChildren = [];
-              for(let k = 0; k < numberOfGreatGrandChildren; k+=1){
-                greatGrandChildren.push({name: ServiceBrances[Math.floor(Math.random()*ServiceBrances.length)]})
-              }
-              grandChild.children = greatGrandChildren;
-            }
-          }
-          return children;
-        })()
+        children: generateChildren(),
       },
     };
   },
@@ -66,26 +40,29 @@ export default {
         .linktree {
           stroke: red !important;
         }
-        ${(()=>{
-          let levelOrderTree = LevelTraverseTree(this.tree);
-          // remove the head of the tree
-          levelOrderTree.shift();
-          // reverse the tree
-          levelOrderTree = levelOrderTree.reverse();
-          let grayCss = '';
-          for(let i = 0; i < levelOrderTree.length; i+=1){
-            if(this.$store.state.checkedServices.includes(levelOrderTree[i]) === false){
-              grayCss += `
-                #my-custom-tree > svg > g > path:nth-child(${i+1}) {
-                  stroke: grey !important;
-                }
-              `;
-            }
-          }
-          return grayCss;
-        })()}
+        ${this.getGrayedOutCss()}
       `;
     }
+  },
+  methods: {
+    getGrayedOutCss(){
+      let levelOrderTree = LevelTraverseTree(this.tree);
+      // remove the head of the tree
+      levelOrderTree.shift();
+      // reverse the tree
+      levelOrderTree = levelOrderTree.reverse();
+      let grayCss = '';
+      for(let i = 0; i < levelOrderTree.length; i+=1){
+        if(this.$store.state.checkedServices.includes(levelOrderTree[i]) === false){
+          grayCss += `
+            #my-custom-tree > svg > g > path:nth-child(${i+1}) {
+              stroke: grey !important;
+            }
+          `;
+        }
+      }
+      return grayCss;
+    },
   },
 }
 </script>
